+++ demo/src/main/java/com/example/demo/repository/EmployeeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.model.Employee;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Long>{
+public interface EmployeeRepository extends JpaRepository<Employee, Integer>{
 
 	Optional<Employee> findByEid(Integer eid);
 }
